# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger API",
        "version": "0.1"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Login user",
                "description": "Login a user",
                "operationId": "loginUser",
                "requestBody": {
                    "description": "User email and password are needed to login",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "$ref": "#/components/schemas/User/properties/email"
                                    },
                                    "password": {
                                        "$ref": "#/components/schemas/User/properties/password"
                                    },
                                    "device_name": {
                                        "type": "string",
                                        "example": "ios"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        },
                                        "name": {
                                            "$ref": "#/components/schemas/User/properties/name"
                                        },
                                        "email": {
                                            "$ref": "#/components/schemas/User/properties/email"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tous les champs sont obligatoires"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Mauvais identifiants"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Register user",
                "description": "Register a user",
                "operationId": "registerUser",
                "requestBody": {
                    "description": "User email, name, password, device_name are needed to register",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "name",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "$ref": "#/components/schemas/User/properties/email"
                                    },
                                    "name": {
                                        "$ref": "#/components/schemas/User/properties/name"
                                    },
                                    "password": {
                                        "$ref": "#/components/schemas/User/properties/password"
                                    },
                                    "device_name": {
                                        "type": "string",
                                        "example": "ios"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        },
                                        "name": {
                                            "$ref": "#/components/schemas/User/properties/name"
                                        },
                                        "email": {
                                            "$ref": "#/components/schemas/User/properties/email"
                                        },
                                        "created_at": {
                                            "$ref": "#/components/schemas/User/properties/created_at"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Veuillez remplir tous les champs ou verifier que votre mot de passe contient bien au moins 8 caractères"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Votre compte existe déjà"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}